Shift recording creation logic to renderer
